+++ strwythura/__init__.py
@@ -6,26 +6,24 @@ Package definitions for Strwythura.
 see copyright/license https://github.com/DerwenAI/strwythura/README.md
 """
 
+from .graph import Entity, TextChunk
+
 from .kg import abstract_overlay, construct_kg
 
-from .lex import STOP_WORDS, \
-    parse_text, \
+from .lex import parse_text, \
     extract_entity, extract_relations, make_entity
 
-from .nlp import NER_LABELS, RE_LABELS, \
+from .nlp import NER_LABELS, RE_LABELS, STOP_WORDS, \
     init_nlp_pipe, make_chunk, uni_scrubber
 
 from .opt import calc_quantile_bins, stripe_column, root_mean_square
 
 from .scrape import scrape_html
 
-from .strw import HTML_PATH, KG_PATH, W2V_PATH, \
-    Strwythura, GraphRAG
+from .strw import Strwythura, GraphRAG
 
 from .textrank import TR_ALPHA, TR_LOOKBACK, \
     run_textrank, cooccur_entities
-
-from .valid import Entity, TextChunk
 
 from .vis import gen_pyvis
